Time: 29 ms (87.26%), Space: 116.7 MB (62.97%) - LeetHub

+++ 973-k-closest-points-to-origin/973-k-closest-points-to-origin.java
@@ -9,7 +9,7 @@ class Solution {
             }
             
             private int distanceToOrigin(int [] point) {
-                return (int) point[0]*point[0] + point[1]*point[1];
+                return point[0]*point[0] + point[1]*point[1];
             }
         };
